Revert "数字を浮動小数点としてトークナイズするように変更"
This reverts commit 282dde01c4e4ed40bea4f20c83acc21a830c77cf.

+++ 9cc.h
@@ -8,7 +8,7 @@ enum {
 
 typedef struct {
   int ty;
-  float val;
+  int val;
   char *input;
 } Token;
